reformat, move tests around
