+++ omen2/errors.py
@@ -9,10 +9,6 @@ class OmenMoreThanOneError(OmenError):
     pass
 
 
-class OmenDuplicateObjectError(OmenError):
-    pass
-
-
 class OmenNoPkError(OmenError, ValueError):
     pass
